+++ administrator/components/com_gamification/controllers/achievement.php
@@ -64,7 +64,6 @@ class GamificationControllerAchievement extends Backend
         // Check for errors
         if ($validData === false) {
             $this->displayNotice($form->getErrors(), $redirectOptions);
-
             return;
         }