+++ administrator/components/com_gamification/views/badge/view.html.php
@@ -30,7 +30,7 @@ class GamificationViewBadge extends JViewLegacy
     protected $documentTitle;
     protected $option;
 
-    protected $imagesFolder;
+    protected $mediaFolder;
     
     public function display($tpl = null)
     {
@@ -42,7 +42,9 @@ class GamificationViewBadge extends JViewLegacy
 
         // Load the component parameters.
         $params             = JComponentHelper::getParams($this->option);
-        $this->imagesFolder = $params->get('images_directory', 'images/gamification');
+        
+        $filesystemHelper   = new Prism\Filesystem\Helper($params);
+        $this->mediaFolder  = $filesystemHelper->getMediaFolder();
 
         // Prepare actions, behaviors, scripts and document
         $this->addToolbar();