+++ administrator/components/com_gamification/views/challenge/tmpl/edit.php
@@ -15,10 +15,10 @@ defined('_JEXEC') or die;
         <form action="<?php echo JRoute::_('index.php?option=com_gamification'); ?>" method="post" name="adminForm" id="adminForm" class="form-validate form-horizontal" enctype="multipart/form-data">
 
             <fieldset class="adminform">
-                <?php echo $this->form->getControlGroup('title'); ?>
-                <?php echo $this->form->getControlGroup('group_id'); ?>
-                <?php echo $this->form->getControlGroup('image'); ?>
-                <?php echo $this->form->getControlGroup('published'); ?>
+                <?php echo $this->form->renderField('title'); ?>
+                <?php echo $this->form->renderField('group_id'); ?>
+                <?php echo $this->form->renderField('image'); ?>
+                <?php echo $this->form->renderField('published'); ?>
 
                 <div class="control-group">
                     <div class="control-label">
@@ -32,8 +32,8 @@ defined('_JEXEC') or die;
                     </div>
                 </div>
 
-                <?php echo $this->form->getControlGroup('note'); ?>
-                <?php echo $this->form->getControlGroup('id'); ?>
+                <?php echo $this->form->renderField('note'); ?>
+                <?php echo $this->form->renderField('id'); ?>
             </fieldset>
 
             <input type="hidden" name="task" value=""/>