+++ administrator/components/com_gamification/views/group/tmpl/edit.php
@@ -15,9 +15,9 @@ defined('_JEXEC') or die;
         <form action="<?php echo JRoute::_('index.php?option=com_gamification'); ?>" method="post" name="adminForm" id="adminForm" class="form-validate form-horizontal">
 
             <fieldset class="adminform">
-                <?php echo $this->form->getControlGroup('name'); ?>
-                <?php echo $this->form->getControlGroup('note'); ?>
-                <?php echo $this->form->getControlGroup('id'); ?>
+                <?php echo $this->form->renderField('name'); ?>
+                <?php echo $this->form->renderField('note'); ?>
+                <?php echo $this->form->renderField('id'); ?>
             </fieldset>
 
             <input type="hidden" name="task" value=""/>