+++ administrator/components/com_gamification/views/level/tmpl/edit.php
@@ -15,14 +15,20 @@ defined('_JEXEC') or die;
         <form action="<?php echo JRoute::_('index.php?option=com_gamification'); ?>" method="post" name="adminForm" id="adminForm" class="form-validate form-horizontal">
 
             <fieldset class="adminform">
-                <?php echo $this->form->getControlGroup('title'); ?>
-                <?php echo $this->form->getControlGroup('group_id'); ?>
-                <?php echo $this->form->getControlGroup('points'); ?>
-                <?php echo $this->form->getControlGroup('points_id'); ?>
-                <?php echo $this->form->getControlGroup('value'); ?>
-                <?php echo $this->form->getControlGroup('rank_id'); ?>
-                <?php echo $this->form->getControlGroup('published'); ?>
-                <?php echo $this->form->getControlGroup('id'); ?>
+                <?php echo $this->form->renderField('title'); ?>
+                <?php echo $this->form->renderField('group_id'); ?>
+                <?php echo $this->form->renderField('value'); ?>
+                <?php echo $this->form->renderField('rank_id'); ?>
+                <?php echo $this->form->renderField('published'); ?>
+                
+                <div class="control-group">
+                    <div class="control-label"><?php echo $this->form->getLabel('points'); ?></div>
+                    <div class="controls">
+                        <?php echo $this->form->getInput('points'); ?>
+                        <?php echo $this->form->getInput('points_id'); ?>
+                    </div>
+                </div>
+                <?php echo $this->form->renderField('id'); ?>
             </fieldset>
 
             <input type="hidden" name="task" value=""/>