+++ administrator/components/com_gamification/views/reward/tmpl/edit.php
@@ -8,50 +8,66 @@
  */
 
 // no direct access
-defined('_JEXEC') or die;
-?>
-<div class="row-fluid">
-    <div class="span8">
-        <form action="<?php echo JRoute::_('index.php?option=com_gamification'); ?>" method="post" name="adminForm" id="adminForm" class="form-validate form-horizontal" enctype="multipart/form-data">
+defined('_JEXEC') or die; ?>
+<form action="<?php echo JRoute::_('index.php?option=com_gamification'); ?>" method="post" name="adminForm" id="adminForm" class="form-validate form-horizontal" enctype="multipart/form-data">
 
+    <?php echo JHtml::_('bootstrap.startTabSet', 'rewardData', array('active' => 'details')); ?>
+
+    <?php echo JHtml::_('bootstrap.addTab', 'rewardData', 'details', JText::_('COM_GAMIFICATION_DETAILS')); ?>
+    <div class="row-fluid">
+        <div class="span8">
             <fieldset class="adminform">
-                <?php echo $this->form->renderField('title'); ?>
-                <?php echo $this->form->renderField('group_id'); ?>
-                <?php echo $this->form->renderField('points'); ?>
-                <?php echo $this->form->renderField('points_id'); ?>
-                <?php echo $this->form->renderField('number'); ?>
-                <?php echo $this->form->renderField('image'); ?>
-                <?php echo $this->form->renderField('published'); ?>
-
-                <div class="control-group">
-                    <div class="control-label">
-                        <?php echo $this->form->getLabel('description'); ?>
-                        <a class="btn btn-mini hasTooltip" href="javascript: void(0);" title="<?php echo JHtml::tooltipText('COM_GAMIFICATION_PLACEHOLDERS', 'COM_GAMIFICATION_PLACEHOLDERS_RANK'); ?>">
-                            <i class="icon-info"></i>
-                        </a>
-                    </div>
-                    <div class="controls">
-                        <?php echo $this->form->getInput('description'); ?>
-                    </div>
-                </div>
-
-                <?php echo $this->form->renderField('note'); ?>
-                <?php echo $this->form->renderField('id'); ?>
+            <?php echo $this->form->renderField('title'); ?>
+            <?php echo $this->form->renderField('group_id'); ?>
+            <?php echo $this->form->renderField('image'); ?>
+            <?php echo $this->form->renderField('resize_image'); ?>
+            <?php echo $this->form->renderField('published'); ?>
+            <?php echo $this->form->renderField('description'); ?>
+            <?php echo $this->form->renderField('id'); ?>
             </fieldset>
-
-            <input type="hidden" name="task" value=""/>
-            <?php echo JHtml::_('form.token'); ?>
-        </form>
+        </div>
+        <?php if ($this->item->image) { ?>
+            <div class="span4">
+                <img src="<?php echo '../' . $this->mediaFolder . '/' . $this->item->image; ?>" class="img-polaroid"/>
+                <br/><br/>
+                <a class="btn btn-danger" href="<?php echo JRoute::_('index.php?option=com_gamification&task=reward.removeImage&id=' . (int)$this->item->id . '&' . JSession::getFormToken() . '=1'); ?>" id="gfy-remove-image">
+                    <i class="icon-trash"></i>
+                    <?php echo JText::_('COM_GAMIFICATION_REMOVE_IMAGE'); ?>
+                </a>
+            </div>
+        <?php } ?>
     </div>
 
-    <?php if ($this->item->image) { ?>
-        <div class="span4">
-            <img src="<?php echo '../' . $this->mediaFolder . '/' . $this->item->image; ?>" class="img-polaroid"/>
-            <br/><br/>
-            <a class="btn btn-danger" href="<?php echo JRoute::_('index.php?option=com_gamification&task=reward.removeImage&id=' . (int)$this->item->id . '&' . JSession::getFormToken() . '=1'); ?>" id="gfy-remove-image">
-                <i class="icon-trash"></i>
-                <?php echo JText::_('COM_GAMIFICATION_REMOVE_IMAGE'); ?>
+    <?php echo JHtml::_('bootstrap.endTab'); ?>
+
+    <?php echo JHtml::_('bootstrap.addTab', 'rewardData', 'advanced', JText::_('JGLOBAL_FIELDSET_ADVANCED')); ?>
+    <div class="control-group">
+    <div class="control-label"><?php echo $this->form->getLabel('points'); ?></div>
+    <div class="controls">
+        <?php echo $this->form->getInput('points'); ?>
+        <?php echo $this->form->getInput('points_id'); ?>
+    </div>
+    </div>
+
+    <?php echo $this->form->renderField('number'); ?>
+
+    <div class="control-group">
+        <div class="control-label">
+            <?php echo $this->form->getLabel('activity_text'); ?>
+            <a class="btn btn-mini hasTooltip" href="javascript: void(0);" title="<?php echo JHtml::tooltipText('COM_GAMIFICATION_PLACEHOLDERS', 'COM_GAMIFICATION_PLACEHOLDERS_DESCRIPTION'); ?>">
+                <i class="icon-info"></i>
             </a>
         </div>
-    <?php } ?>
-</div>
+        <div class="controls">
+            <?php echo $this->form->getInput('activity_text'); ?>
+        </div>
+    </div>
+
+    <?php echo $this->form->renderField('note'); ?>
+
+    <?php echo JHtml::_('bootstrap.endTab'); ?>
+    <?php echo JHtml::_('bootstrap.endTabSet'); ?>
+
+    <input type="hidden" name="task" value=""/>
+    <?php echo JHtml::_('form.token'); ?>
+</form>